fix require cycles
