+++ playground/datepicker/DatePicker.tsx
@@ -1,30 +1,16 @@
-import React, {createContext, useReducer, useContext, useState, Dispatch} from 'react';
+import React, {useReducer, useState } from 'react';
 import {View, StyleSheet, StyleProp, ViewStyle} from 'react-native';
 import PropTypes from 'prop-types';
-
-import {Calendar, SelectMonth, SelectTime} from './components';
-import {Utils} from './Utils';
-import {Configs, Mode} from './types';
-
-const defaultOptions: Options = {
-  backgroundColor: '#fff',
-  textHeaderColor: '#212c35',
-  textDefaultColor: '#2d4150',
-  selectedTextColor: '#fff',
-  mainColor: '#61dafb',
-  textSecondaryColor: '#7a92a5',
-  borderColor: 'rgba(122, 146, 165, 0.1)',
-  defaultFont: 'System',
-  headerFont: 'System',
-  textFontSize: 15,
-  textHeaderFontSize: 17,
-  headerAnimationDistance: 100,
-  daysAnimationDistance: 200,
-};
+import Calendar from './components/Calendar';
+import SelectMonth from './components/SelectMonth';
+import SelectTime from './components/SelectTime';
+import Utils from './Utils';
+import type {Configs, Mode} from './types';
+import CalendarContext, { type CalendarContextType, defaultOptions } from './context/CalendarContext';
 
 type ActionType = 'set' | 'toggleMonth' | 'toggleTime';
 
-type Action = {
+export type Action = {
   type: ActionType;
   selectedDate?: string;
   activeDate?: string;
@@ -41,57 +27,6 @@ const reducer = (state: CalendarState, action: Action): CalendarState => {
     default:
       throw new Error('Unexpected action');
   }
-};
-
-type CalendarContextType = {
-  mode: Mode;
-  reverse: boolean;
-  options: Options;
-  utils: Utils;
-  state: [CalendarState, Dispatch<Action>];
-  onSelectedChange?: (selectedDate: string) => void;
-  onDateChange?: (formattedDate: string) => void;
-  disableDateChange: boolean;
-  minimumDate: string;
-  maximumDate: string;
-  selectorStartingYear: number;
-  selectorEndingYear: number;
-  onMonthYearChange?: (formattedDate: string) => void;
-  minuteInterval: MinuteInterval;
-  onTimeChange?: (time: string) => void;
-};
-
-const CalendarContext = createContext<CalendarContextType>({
-  mode: 'datepicker',
-  reverse: false,
-  options: defaultOptions,
-  utils: new Utils({
-    minimumDate: '',
-    maximumDate: '',
-    configs: {},
-    reverse: false,
-    mode: 'datepicker',
-  }),
-  state: [
-    {
-      activeDate: '',
-      selectedDate: '',
-      monthOpen: false,
-      timeOpen: false,
-    },
-    () => null,
-  ],
-  disableDateChange: false,
-  minimumDate: '',
-  maximumDate: '',
-  selectorStartingYear: 0,
-  selectorEndingYear: 0,
-  minuteInterval: 5,
-});
-
-const useCalendar = () => {
-  const contextValue = useContext(CalendarContext);
-  return contextValue;
 };
 
 export type Options = {
@@ -130,7 +65,7 @@ type Props = {
   style?: StyleProp<ViewStyle>;
 };
 
-type CalendarState = {
+export type CalendarState = {
   activeDate: string;
   selectedDate: string;
   monthOpen: boolean;
@@ -252,7 +187,7 @@ const optionsShape = {
 };
 const modeArray: Array<Mode> = ['datepicker', 'calendar', 'monthYear', 'time'];
 const minuteIntervalArray = [1, 2, 3, 4, 5, 6, 10, 12, 15, 20, 30, 60] as const;
-type MinuteInterval = typeof minuteIntervalArray[number];
+export type MinuteInterval = typeof minuteIntervalArray[number];
 
 DatePicker.defaultProps = {
   onSelectedChange: () => null,
@@ -294,4 +229,4 @@ DatePicker.propTypes = {
   style: PropTypes.oneOfType([PropTypes.object, PropTypes.array]),
 };
 
-export {DatePicker, CalendarContext, useCalendar};
+export default DatePicker;
