--- playground/datepicker/components/index.ts
@@ -1,5 +1,0 @@
-export * from './Header';
-export * from './Days';
-export * from './SelectMonth';
-export * from './Calendar';
-export * from './SelectTime';
