+++ playground/datepicker/hooks/useCalendar.ts
@@ -1,0 +1,9 @@
+import {useContext} from 'react';
+import CalendarContext from '../context/CalendarContext';
+
+const useCalendar = () => {
+  const contextValue = useContext(CalendarContext);
+  return contextValue;
+};
+
+export default useCalendar;
