+++ playground/datepicker/index.ts
@@ -1,5 +1,5 @@
-import {DatePicker} from './DatePicker';
-import {Utils} from './Utils';
+import DatePicker from './DatePicker';
+import Utils from './Utils';
 
 export const {getFormatedDate, getToday} = new Utils({});
 export default DatePicker;
